Rename models and renderers to match Minecraft; comment out animation controller registration for conch on head
