+++ src/main/java/com/fredtargaryen/rocketsquids/client/render/BabyRocketSquidRenderer.java
@@ -3,7 +3,7 @@
 package com.fredtargaryen.rocketsquids.client.render;
 
 import com.fredtargaryen.rocketsquids.DataReference;
-import com.fredtargaryen.rocketsquids.client.model.ModelRocketSquidBaby;
+import com.fredtargaryen.rocketsquids.client.model.BabyRocketSquidModel;
 import com.fredtargaryen.rocketsquids.level.entity.BabyRocketSquidEntity;
 import com.mojang.blaze3d.vertex.PoseStack;
 import com.mojang.math.Axis;
@@ -16,13 +16,13 @@ import org.jetbrains.annotations.NotNull;
 import static com.fredtargaryen.rocketsquids.client.event.ClientHandler.BABY_SQUID_BODY_LAYER;
 
 @SuppressWarnings("removal")
-public class RenderBabyRS extends MobRenderer<BabyRocketSquidEntity, ModelRocketSquidBaby<BabyRocketSquidEntity>> {
+public class BabyRocketSquidRenderer extends MobRenderer<BabyRocketSquidEntity, BabyRocketSquidModel<BabyRocketSquidEntity>> {
     private static final ResourceLocation normal = new ResourceLocation(DataReference.MODID + ":textures/entity/baby_rocket_squid.png");
 
-    public RenderBabyRS(
+    public BabyRocketSquidRenderer(
             EntityRendererProvider.Context context
     ) {
-        super(context, new ModelRocketSquidBaby<>(context.bakeLayer(BABY_SQUID_BODY_LAYER)), 1.0f);
+        super(context, new BabyRocketSquidModel<>(context.bakeLayer(BABY_SQUID_BODY_LAYER)), 1.0f);
     }
 
     /**
