+++ src/main/java/com/fredtargaryen/rocketsquids/level/item/GeoModArmorItem.java
@@ -2,17 +2,11 @@
 // See README.md for full copyright notice and contributor info
 package com.fredtargaryen.rocketsquids.level.item;
 
-import net.minecraft.world.entity.Entity;
-import net.minecraft.world.entity.decoration.ArmorStand;
 import net.minecraft.world.item.ArmorItem;
 import net.minecraft.world.item.ArmorMaterial;
 import software.bernie.geckolib.animatable.GeoItem;
-import software.bernie.geckolib.constant.DataTickets;
-import software.bernie.geckolib.constant.DefaultAnimations;
 import software.bernie.geckolib.core.animatable.instance.AnimatableInstanceCache;
 import software.bernie.geckolib.core.animation.AnimatableManager;
-import software.bernie.geckolib.core.animation.AnimationController;
-import software.bernie.geckolib.core.object.PlayState;
 import software.bernie.geckolib.util.GeckoLibUtil;
 
 public class GeoModArmorItem extends ArmorItem implements GeoItem {
@@ -25,23 +19,23 @@ public class GeoModArmorItem extends ArmorItem implements GeoItem {
     // Create our animation controller
     @Override
     public void registerControllers(AnimatableManager.ControllerRegistrar controllers) {
-        controllers.add(new AnimationController<>(this, 20, state -> {
-            // Apply our generic idle animation.
-            // When it plays is decided below
-            state.setAnimation(DefaultAnimations.IDLE);
-
-            // Gather some information about the entity wearing it
-            Entity entity = state.getData(DataTickets.ENTITY);
-
-            // Check if it is an armor stand
-            if (entity instanceof ArmorStand) {
-                // if so don't play
-                return PlayState.STOP;
-            } else {
-                // if not do play
-                return PlayState.CONTINUE;
-            }
-        }));
+//        controllers.add(new AnimationController<>(this, 20, state -> {
+//            // Apply our generic idle animation.
+//            // When it plays is decided below
+//            state.setAnimation(DefaultAnimations.IDLE);
+//
+//            // Gather some information about the entity wearing it
+//            Entity entity = state.getData(DataTickets.ENTITY);
+//
+//            // Check if it is an armor stand
+//            if (entity instanceof ArmorStand) {
+//                // if so don't play
+//                return PlayState.STOP;
+//            } else {
+//                // if not do play
+//                return PlayState.CONTINUE;
+//            }
+//        }));
     }
 
     @Override
